ExpenseTracker/dsserviceRegEx/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
kafka-python==2.0.2
python-dotenv==1.0.1
